requirements.txt
# requirements.txt (clean + stable)

# LangChain ecosystem
langchain==0.3.26
langchain-openai>=1.86.0,<2.0.0
langchain-community==0.3.26
langchain-pinecone==0.2.0

# Vector DB + Embeddings
pinecone-client==5.0.0
sentence-transformers==3.0.1
faiss-cpu==1.7.4

# PDF processing
pypdf==5.6.1
pdfplumber==0.9.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
numpy==1.26.4
pandas==2.2.2

# NLP
spacy==3.8.11
nltk==3.9.1

# Web / API Servers
streamlit==1.39.0
fastapi==0.115.0
uvicorn==0.30.1
flask==3.1.0

# Monitoring
prometheus_client==0.20.0

# Testing
pytest==8.3.3
